Restructure learning path pages. Add tsconfig path mappings
