+++ frontends/mit-open/src/page-components/LearningResourceCard/LearningResourceCard.test.tsx
@@ -1,14 +1,14 @@
 import React from "react"
 import * as NiceModal from "@ebay/nice-modal-react"
 import { makeLearningResource } from "ol-search-ui/src/test-utils/factories"
-import { renderWithProviders, user, screen } from "../test-utils"
-import type { User } from "../test-utils"
+import { renderWithProviders, user, screen } from "../../test-utils"
+import type { User } from "../../test-utils"
 import LearningResourceCard from "./LearningResourceCard"
 import type {
   LearningResourceCardPropsOld,
   LearningResourceCardPropsNew,
 } from "./LearningResourceCard"
-import AddToListDialog from "../pages/learningpaths/AddToListDialog"
+import AddToListDialog from "./AddToListDialog"
 import * as factories from "api/test-utils/factories"
 
 jest.mock("@ebay/nice-modal-react", () => {
