+++ frontends/mit-open/src/pages/LearningPathDetailsPage/ItemsListing.test.tsx
@@ -1,7 +1,7 @@
 import React from "react"
 import { faker } from "@faker-js/faker/locale/en"
 import { SortableList, SortableItem } from "ol-util"
-import LearningResourceCard from "../../components/LearningResourceCard"
+import LearningResourceCard from "page-components/LearningResourceCard/LearningResourceCard"
 import { factories, urls, makeRequest } from "api/test-utils"
 import {
   screen,
