+++ frontends/mit-open/src/pages/LearningPathDetailsPage/ItemsListing.tsx
@@ -1,7 +1,7 @@
 import React, { useCallback, useEffect } from "react"
 import classNames from "classnames"
 import type { LearningPathRelationship } from "api"
-import LearningResourceCard from "../../components/LearningResourceCard"
+import LearningResourceCard from "page-components/LearningResourceCard/LearningResourceCard"
 import {
   SortableItem,
   SortableList,
