+++ frontends/mit-open/src/routes.tsx
@@ -2,8 +2,8 @@ import React from "react"
 import { RouteObject, Outlet } from "react-router"
 import HomePage from "./pages/HomePage/HomePage"
 import RestrictedRoute from "./components/RestrictedRoute"
-import LearningPathListingPage from "./pages/learningpaths/LearningPathListingPage"
-import LearningPathDetailsPage from "./pages/learningpaths/LearningPathDetails"
+import LearningPathListingPage from "./pages/LearningPathListingPage/LearningPathListingPage"
+import LearningPathDetailsPage from "./pages/LearningPathDetailsPage/LearningPathDetailsPage"
 import ArticleDetailsPage from "./pages/articles/ArticleDetailsPage"
 import {
   ArticlesCreatePage,
