+++ frontends/mit-open/tsconfig.json
@@ -2,7 +2,12 @@
   "extends": "../../tsconfig.json",
   "compilerOptions": {
     "outDir": "./build",
-    "rootDir": "./src"
+    "rootDir": "./src",
+    "baseUrl": "./src",
+    "paths": {
+      "page-components/*": ["page-components/*"],
+      "components/*": ["components/*"]
+    }
   },
   "include": ["./src/**/*.ts", "./src/**/*.tsx"],
   "typeRoots": ["./src/typings", "./node_modules/@types/"]
